AES-8
Die Datei:
	private int idUser;
	private String nameForClub;
	private String gender;
	private Date birthdate;
	private String country;
	private String city;
	private String denomination;

	wurde erfolgreich im Datenbank gespeichert.

+++ src/main/java/com/aeClub/util/SecurityUtil.java
@@ -22,7 +22,7 @@ public final class SecurityUtil {
 		}
 	}
 
-	public static Long getCurrentIdProfile() {
+	public static int getCurrentIdProfile() {
 		CurrentProfile currentProfile = getCurrentProfile();
 		return currentProfile != null ? currentProfile.getId() : 0;
 	}
